requirements.txt
django
pandas
psycopg2
djangorestframework
python-dotenv
pyarrow
pysus